move issue21 test to test_issues and run it
